Refactor interactions to own file.
